+++ cypress/integration/browser.spec.ts
@@ -1,0 +1,4 @@
+window.ENV = { development: true, production: false, node: false };
+
+import '../../src/analytics/test/browser/result_collector.spec';
+import '../../src/input/test/browser/exif_helper.spec';
