+++ cypress/integration/universal.spec.ts
@@ -1,0 +1,10 @@
+window.ENV = { development: true, production: false, node: false };
+
+import '../../src/common/test/array_helper.spec';
+import '../../src/common/test/cv_utils.spec';
+import '../../src/common/test/events.spec';
+import '../../src/locator/test/barcode_locator.spec';
+import '../../src/quagga/test/transform.spec';
+
+// import '../../test/spec/camera_access.spec';
+// import '../../test/integration/integration.spec';
