+++ package.json
@@ -15,6 +15,9 @@
     "@babel/preset-env": "^7.9.6",
     "@babel/preset-typescript": "^7.8.3",
     "@babel/runtime": "^7.9.6",
+    "@cypress/code-coverage": "^3.7.0",
+    "@cypress/webpack-preprocessor": "^5.1.2",
+    "@std/esm": "^0.26.0",
     "@types/chai": "^4.2.9",
     "@types/gl-vec2": "^1.3.0",
     "@types/lodash": "^4.14.150",
@@ -28,17 +31,18 @@
     "chai": "^4.2.0",
     "core-js": "^3.6.5",
     "cross-env": "^7.0.0",
+    "cypress": "^4.5.0",
     "eslint": "^6.8.0",
     "grunt": "^1.0.4",
     "grunt-cli": "^1.3.2",
     "grunt-contrib-nodeunit": "1.0.0",
     "grunt-karma": "2.0.0",
     "isparta-loader": "^2.0.0",
+    "istanbul-lib-coverage": "^3.0.0",
     "karma": "1.3.0",
     "karma-chai": "0.1.0",
     "karma-chrome-launcher": "2.2.0",
     "karma-coverage": "1.1.2",
-    "karma-firefox-launcher": "0.1.7",
     "karma-mocha": "1.3.0",
     "karma-phantomjs-launcher": "0.2.3",
     "karma-sinon": "^1.0.5",
@@ -46,10 +50,13 @@
     "karma-source-map-support": "^1.4.0",
     "karma-webpack": "^4.0.2",
     "mocha": "5.2.0",
+    "nyc": "^15.0.1",
     "phantomjs": "^2.1.7",
     "sinon": "5.1.1",
     "sinon-chai": "^3.5.0",
     "source-map-loader": "^0.2.4",
+    "ts-mocha": "^7.0.0",
+    "ts-node": "^8.10.1",
     "typescript": "^3.8.2",
     "webpack": "^4.43.0",
     "webpack-cli": "^3.3.11",
@@ -60,13 +67,21 @@
   },
   "scripts": {
     "check-types": "tsc",
-    "test": "npx grunt test",
-    "test:require": "npx mocha test/test-require",
-    "test:import": "npx mocha -r @std/esm -r ts-node/register test/test-import.mjs",
+    "coverage": "npm run cypress:run && npx cross-env NODE_ENV=test BUILD_ENV=development nyc --no-clean ts-mocha -p test/tsconfig.json src/**/test/node/*.spec.* src/**/test/*.spec.*",
+    "cypress:open": "npx cross-env NODE_ENV=development BUILD_ENV=development cypress open --env BUILD_ENV=development",
+    "cypress:run": "npx cross-env NODE_ENV=test BUILD_ENV=development cypress run --env BUILD_ENV=development",
+    "test:browser-specific": "NOT WORKING -- something like npx cypress run --config testFiles=[browser]",
+    "test:browser-universal": "NOT WORKING -- something like npx cypress run --config testFiles=[universal]",
+    "test:browser-all": "npm run cypress:run",
+    "test:import": "mocha -r @std/esm test/test-import.mjs",
+    "test:node": "npx cross-env NODE_ENV=test ts-mocha -p test/tsconfig.json src/**/test/node/*.spec.* src/**/test/*.spec.*",
+    "test:require": "mocha test/test-require",
+    "test:module": "npm run test:require && npm run test:import",
+    "test": "npm run cypress:run && npx cross-env NODE_ENV=test BUILD_ENV=development ts-mocha -p test/tsconfig.json src/**/test/node/*.spec.* src/**/test/*.spec.* && npx cross-env NODE_ENV=production BUILD_ENV=production grunt integrationtest",
     "integrationtest": "grunt integrationtest",
-    "build:dev": "npx cross-env BUILD_ENV=development webpack --config configs/webpack.config.js",
-    "build:prod": "npx cross-env BUILD_ENV=production webpack --config configs/webpack.config.min.js",
-    "build:node": "npx cross-env BABEL_ENV=commonjs BUILD_ENV=node webpack --config configs/webpack.node.config.js",
+    "build:dev": "npx cross-env NODE_ENV=development BUILD_ENV=development webpack --config configs/webpack.config.js",
+    "build:prod": "npx cross-env NODE_ENV=production BUILD_ENV=production webpack --config configs/webpack.config.min.js",
+    "build:node": "npx cross-env NODE_ENV=production BABEL_ENV=commonjs BUILD_ENV=node webpack --config configs/webpack.node.config.js",
     "build": "npm run check-types && npm run build:dev && npm run build:prod && npm run build:node",
     "watch": "npx cross-env BUILD_ENV=development webpack --watch",
     "lint": "eslint src",
@@ -138,6 +153,33 @@
     "snyk": "^1.315.1"
   },
   "snyk": true,
+  "nyc": {
+    "include": [
+      "src/**/*.ts",
+      "src/**/*.tsx",
+      "src/**/*.js",
+      "src/**/*.jsx"
+    ],
+    "exclude": [
+      "**/*.spec.*"
+    ],
+    "all": true,
+    "extension": [
+      ".js",
+      ".jsx",
+      ".ts",
+      ".tsx"
+    ],
+    "require": [
+      "ts-node/register"
+    ],
+    "reporter": [
+      "text-summary",
+      "html"
+    ],
+    "sourceMap": true,
+    "instrument": true
+  },
   "greenkeeper": {
     "ignore": [
       "async",
